docs: update README to even remotely reflect project status
the top-level README was initially a tracker for the project back as it was initially conceived.
that was long ago deprecated, for various reasons, but the README remained there frozen in amber.
I've updated it to reflect more of the completed tasks in the intervening time, and moved around
some objectives from the various milestones. milestones 2 and 3 may indeed be intractable at this
point, but in case development on this project continues beyond the very near future, I've left them
there to suggest some directions