chore: remove unmaintained plink support
this plink pipeline was an old relic of a previous version of the pipeline. it's not been maintained
of late and is not safe for use. i think it's better to just remove it to make it much less likely
that people will just stumble in and think plink is usable as is

re #6