fix: update makefile include directives to more conclusively include the right thing
with some recent changes trying to support multiple instances of the pipeline communicating with one
another (specifically so people can skip preprocessing on the same platform), a weird discrepancy
was introduced that caused subworkflows to pull in the wrong config file in some instances, breaking
everything horribly lol. now the top-level makefile reports exactly which copy of Makefile.config it
loads, and that exact file is reloaded to deal with make's bizarre expansion issues with
subworkflows.